refactor(core): extract shared action-sheet helpers to core, fix at source
The action-sheet close + positioning bugs lived in helpers shared by Remove-from-
Continue-Watching, multi-select Remove and arr Search — so fixing them only in the
arr Search path was a cheap local fix that left the same bugs in Remove. Move the
shared helpers (buildNativeActionSheetItem, setActionSheetItemIcon,
getActiveActionSheetScroller, fitActionSheetItem, closeOpenActionSheet) out of the
remove-home feature file into core/action-sheet.ts — one source — and fix both bugs
there so every injector benefits:

- closeOpenActionSheet: the v12 sheet is a dialogHelper <div> (NOT a <dialog>), so
  dialog.close()/Escape were no-ops; replay the outside-tap (mousedown+click on the
  .dialogContainer) → dialogHelper's real close(), guarded history.back() fallback.
- fitActionSheetItem (was fitRemoveItemToMenu): now corrects BOTH the right-edge and
  bottom-edge overflow an appended item causes.

Removes the arr Search duplicate close/fit, and the arr→enhanced/features cross-area
import. remove-home + remove-multiselect + arr Search all import from core now.
